+++ scent.py
@@ -39,7 +39,7 @@ def python(*_):
         (('make', 'test-all'), "Combined Tests", False),
         (('make', 'check'), "Static Analysis", True),
         (('make', 'doc'), None, True),
-        (('make', 'doctor', '-C', 'examples'), "Examples", True),
+        (('make', 'demo'), "Run Example", True),
     ), start=1):
 
         if not run(command, title, count, retry):